+++ libs/Core/Model.php
@@ -20,6 +20,15 @@ abstract class Model
     // or delete the field name of the soft delete
     protected static $softdelete = '';
 
+    // Define the field name of the create time, if it is empty, it means that the create time is not enabled
+    protected static $createTime = '';
+
+    // Define the field name of the update time, if it is empty, it means that the update time is not enabled
+    protected static $updateTime = '';
+
+    // Define the field name of the delete time, if it is empty, it means that the delete time is not enabled
+    protected static $deleteTime = '';
+
     public static function get(array $where = [], array $field = [])
     {
         try {
@@ -41,6 +50,8 @@ abstract class Model
         return $result;
     }
 
+
+
     public static function all($where = [], array $field = [])
     {
         return static::getAll($where, $field);
@@ -56,10 +67,30 @@ abstract class Model
         return $result;
     }
 
+    public static function insertAll(array $data)
+    {
+        try {
+            $result = DB::link(static::$connection)->table(static::tablename())->insertAll($data);
+        } catch (\Exception $e) {
+            return $e->getMessage();
+        }
+        return $result;
+    }
+
     public static function update(array $where, array $data)
     {
         try {
             $result = DB::link(static::$connection)->table(static::tablename())->where($where)->update($data);
+        } catch (\Exception $e) {
+            return $e->getMessage();
+        }
+        return $result;
+    }
+
+    public static function updateAll(array $data)
+    {
+        try {
+            $result = DB::link(static::$connection)->table(static::tablename())->updateAll($data);
         } catch (\Exception $e) {
             return $e->getMessage();
         }
@@ -81,6 +112,11 @@ abstract class Model
         return $result;
     }
 
+    public static function save(array $where, array $data)
+    {
+        return static::updateOrInsert($where, $data);
+    }
+
     public static function delete(array $where = [])
     {
         try {
@@ -95,10 +131,44 @@ abstract class Model
         return $result;
     }
 
+    public static function deleteAll()
+    {
+        try {
+            if (empty(static::$softdelete)) {
+                $result = DB::link(static::$connection)->table(static::tablename())->deleteAll();
+            } else {
+                $result = DB::link(static::$connection)->table(static::tablename())->updateAll([static::$softdelete => 1]);
+            }
+        } catch (\Exception $e) {
+            return $e->getMessage();
+        }
+        return $result;
+    }
+
+    public static function truncate()
+    {
+        try {
+            $result = DB::link(static::$connection)->table(static::tablename())->truncate();
+        } catch (\Exception $e) {
+            return $e->getMessage();
+        }
+        return $result;
+    }
+
     public static function count(array $where = [])
     {
         try {
             $result = DB::link(static::$connection)->table(static::tablename())->where(static::handleWhere($where))->count();
+        } catch (\Exception $e) {
+            return $e->getMessage();
+        }
+        return $result;
+    }
+
+    public static function sum($field, array $where = [])
+    {
+        try {
+            $result = DB::link(static::$connection)->table(static::tablename())->where(static::handleWhere($where))->sum($field);
         } catch (\Exception $e) {
             return $e->getMessage();
         }
@@ -160,6 +230,126 @@ abstract class Model
     {
         try {
             $result = DB::link(static::$connection)->table(static::tablename())->where(static::handleWhere($where))->getSql();
+        } catch (\Exception $e) {
+            return $e->getMessage();
+        }
+        return $result;
+    }
+
+    public static function query(string $sql)
+    {
+        try {
+            $result = DB::link(static::$connection)->query($sql);
+        } catch (\Exception $e) {
+            return $e->getMessage();
+        }
+        return $result;
+    }
+
+    public static function execute(string $sql)
+    {
+        try {
+            $result = DB::link(static::$connection)->execute($sql);
+        } catch (\Exception $e) {
+            return $e->getMessage();
+        }
+        return $result;
+    }
+
+    public static function beginTransaction()
+    {
+        try {
+            $result = DB::link(static::$connection)->beginTransaction();
+        } catch (\Exception $e) {
+            return $e->getMessage();
+        }
+        return $result;
+    }
+
+    public static function commit()
+    {
+        try {
+            $result = DB::link(static::$connection)->commit();
+        } catch (\Exception $e) {
+            return $e->getMessage();
+        }
+        return $result;
+    }
+
+    public static function rollback()
+    {
+        try {
+            $result = DB::link(static::$connection)->rollback();
+        } catch (\Exception $e) {
+            return $e->getMessage();
+        }
+        return $result;
+    }
+
+    public static function getFields()
+    {
+        try {
+            $result = DB::link(static::$connection)->getFields(static::tablename());
+        } catch (\Exception $e) {
+            return $e->getMessage();
+        }
+        return $result;
+    }
+
+    public static function getPrimary()
+    {
+        try {
+            $result = DB::link(static::$connection)->getPrimaryKey(static::tablename());
+        } catch (\Exception $e) {
+            return $e->getMessage();
+        }
+        return $result;
+    }
+
+    public static function getTable()
+    {
+        try {
+            $result = DB::link(static::$connection)->getTable(static::tablename());
+        } catch (\Exception $e) {
+            return $e->getMessage();
+        }
+        return $result;
+    }
+
+    public static function getTableInfo()
+    {
+        try {
+            $result = DB::link(static::$connection)->getTableInfo(static::tablename());
+        } catch (\Exception $e) {
+            return $e->getMessage();
+        }
+        return $result;
+    }
+
+    public static function getTableStatus()
+    {
+        try {
+            $result = DB::link(static::$connection)->getTableStatus(static::tablename());
+        } catch (\Exception $e) {
+            return $e->getMessage();
+        }
+        return $result;
+    }
+
+    public static function getTables()
+    {
+        try {
+            $result = DB::link(static::$connection)->getTables();
+        } catch (\Exception $e) {
+            return $e->getMessage();
+        }
+        return $result;
+    }
+
+    public static function getDatabases()
+    {
+        try {
+            $result = DB::link(static::$connection)->getDatabases();
         } catch (\Exception $e) {
             return $e->getMessage();
         }